Nightly Build June 4, 2015
Polished current GUI items.
Added world package for playing in the world.
Added gamestates to handle playing/pausing/etc.
Added player package for Characters and NPCs.

Most world and map date will be contained in the TiledMaps data or in
text files that can be loaded separately. This can be discussed later.