Bag
Added a loot bag that should spawn in an enemy's previous location when
an enemy is defeated as well as a "H" version for when the player is
next to the bag and it has not been opened. Bags should de-spawn once
emptied and for this reason there is no "O" version.