AttackAni1
Added attack animations. They SS is size 32 width 64 height. Sprite 1
and 2 are an idle ani that should be played while char is idle. Sprite 3
and 4 are the attack ani. When the char attacks he should 'teleport'
forward first from idle animation to animation 3. Then he should stay in
the same place, go into sprite 4 and then teleport back. Sprite 5 is the
blur from the weapon. It should be displayed statically when sprite 4 is
displayed. the next sprites are only for refrence as to where sprite 5
should go. I made the heights the same for convenience. Time to make the
drinking ani now