Added Friend Movement and Attack SS
It may be tough, but once friend joins party, it's be cool if he could
follow the player around and go to the player's previous location. The
algorithm that would make that happen would be tough and we can say that
that took a majority of our time. I'll be adding an image of Friend that
is highlighted like a chest and we can place him in the game like a
lootable object, but the player can 'talk' instead of looting. Then we
can have to tool tip pop up with the text handler and have some small
paragraph that ends with him joining your party. Maybe we can add a
party management system? idk just spit balling ideas. Also he should
fight with the player.